+++ collector/config.go
@@ -8,6 +8,7 @@ import (
 	"github.com/godror/godror/dsn"
 	"github.com/oracle/oracle-db-appdev-monitoring/azvault"
 	"github.com/oracle/oracle-db-appdev-monitoring/ocivault"
+	"github.com/prometheus/exporter-toolkit/web"
 	"gopkg.in/yaml.v2"
 	"log/slog"
 	"os"
@@ -16,10 +17,18 @@ import (
 )
 
 type MetricsConfiguration struct {
-	MetricsPath string                    `yaml:"metricsPath"`
-	Databases   map[string]DatabaseConfig `yaml:"databases"`
-	Metrics     MetricsFilesConfig        `yaml:"metrics"`
-	Logging     LoggingConfig             `yaml:"log"`
+	ListenAddress string                    `yaml:"listenAddress"`
+	MetricsPath   string                    `yaml:"metricsPath"`
+	Databases     map[string]DatabaseConfig `yaml:"databases"`
+	Metrics       MetricsFilesConfig        `yaml:"metrics"`
+	Logging       LoggingConfig             `yaml:"log"`
+	Web           WebConfig                 `yaml:"web"`
+}
+
+type WebConfig struct {
+	ListenAddresses *[]string `yaml:"listenAddresses"`
+	SystemdSocket   *bool     `yaml:"systemdSocket"`
+	ConfigFile      *string   `yaml:"configFile"`
 }
 
 type DatabaseConfig struct {
@@ -173,7 +182,7 @@ func (d DatabaseConfig) isAzureVault() bool {
 	return d.Vault != nil && d.Vault.Azure != nil
 }
 
-func LoadMetricsConfiguration(logger *slog.Logger, cfg *Config, path string) (*MetricsConfiguration, error) {
+func LoadMetricsConfiguration(logger *slog.Logger, cfg *Config, path string, flags *web.FlagConfig) (*MetricsConfiguration, error) {
 	m := &MetricsConfiguration{}
 	if len(cfg.ConfigFile) > 0 {
 		content, err := os.ReadFile(cfg.ConfigFile)
@@ -196,18 +205,39 @@ func LoadMetricsConfiguration(logger *slog.Logger, cfg *Config, path string) (*M
 		m.Databases["default"] = m.defaultDatabase(cfg)
 	}
 
-	m.merge(cfg, path)
+	m.merge(cfg, path, flags)
 	return m, nil
 }
 
-func (m *MetricsConfiguration) merge(cfg *Config, path string) {
+func (wc WebConfig) Flags() *web.FlagConfig {
+	return &web.FlagConfig{
+		WebListenAddresses: wc.ListenAddresses,
+		WebSystemdSocket:   wc.SystemdSocket,
+		WebConfigFile:      wc.ConfigFile,
+	}
+}
+
+func (m *MetricsConfiguration) merge(cfg *Config, path string, flags *web.FlagConfig) {
 	if len(m.MetricsPath) == 0 {
 		m.MetricsPath = path
 	}
+	m.mergeWebConfig(flags)
 	m.mergeLoggingConfig(cfg)
 	m.mergeMetricsConfig(cfg)
 	if m.Metrics.ScrapeInterval == nil {
 		m.Metrics.ScrapeInterval = &cfg.ScrapeInterval
+	}
+}
+
+func (m *MetricsConfiguration) mergeWebConfig(flags *web.FlagConfig) {
+	if m.Web.ListenAddresses == nil {
+		m.Web.ListenAddresses = flags.WebListenAddresses
+	}
+	if m.Web.SystemdSocket == nil {
+		m.Web.SystemdSocket = flags.WebSystemdSocket
+	}
+	if m.Web.ConfigFile == nil {
+		m.Web.ConfigFile = flags.WebConfigFile
 	}
 }