+++ changelog.md
@@ -5,6 +5,7 @@
 Our current priorities are support for Exadata metrics.  We expect to address these in an upcoming release.
 
 This release includes the following changes:
+- Enable configuration of the prometheus webserver from the config file using the `web` prefix.
 - Allow loading of database password(s) from a file.
 - Fixed a bug where database type (CDB, PDB, etc.) was not reported in certain situations.
 - Fixed a bug where literal passwords containing the '$' character (in the config file) would be evaluated as environment variables. To use literal passwords with the '$' character, escape the '$' character with a second '$': `$test$pwd` becomes `$$test$$pwd`.